pedigree
2016/12/1

+++ src/com/pedigree/R/Diag.java
@@ -2,9 +2,9 @@ package com.pedigree.R;
 
 public class Diag {
 	
-	private static int[][] array;
-	public static int[][] diag(int n){
-		array = new int[n][n];
+	private static double[][] array;
+	public static double[][] diag(int n){
+		array = new double[n][n];
 		for(int i=0; i<n; i++){
 			array[i][i]=1;
 		}
